+++ Sourcecode/Song.ViewData/Methods/TestPaper.cs
@@ -114,7 +114,7 @@ namespace Song.ViewData.Methods
                             WeiSha.Core.Upload.Get[PathKey].DeleteFile(old.Tp_Logo);
                     }
                     //如果没有上传图片，且新对象没有图片，则删除旧图
-                    else if (string.IsNullOrWhiteSpace(old.Tp_Logo))
+                    else if (string.IsNullOrWhiteSpace(entity.Tp_Logo) && !string.IsNullOrWhiteSpace(old.Tp_Logo))
                     {
                         WeiSha.Core.Upload.Get[PathKey].DeleteFile(old.Tp_Logo);
                     }
